pyproject.toml
[tool.poetry]
name = "brn-kb-review"
version = "0.1.0"
description = "Broadnet KB Review Service - Automated knowledge base quality analysis"
authors = ["Broadnet"]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = ">=0.131.0"
uvicorn = ">=0.34.0"
pydantic = ">=2.12.0"
pydantic-settings = ">=2.13.0"
python-dotenv = ">=1.2.0"
langchain-core = ">=0.3.21"
langchain-openai = ">=0.3.0"
langchain-anthropic = ">=1.4.0"
neo4j = ">=5.27.0"
mysql-connector-python = ">=9.2.0"
sqlalchemy = ">=2.0.46,<2.1"
numpy = ">=2.0"
scikit-learn = ">=1.6.0"
PyJWT = ">=2.8.0"
redis = ">=5.0.0"

[tool.poetry.group.test.dependencies]
pytest = ">=9.0.0"
pytest-asyncio = ">=1.0.0"
httpx = ">=0.28.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
